+++ lib/firebase.js
@@ -1,5 +1,5 @@
 import { initializeApp } from 'firebase/app';
-import { child, get, getDatabase, ref } from 'firebase/database';
+import { child, get, getDatabase, ref, set, remove, push } from "firebase/database";
 import decode from '../utils/decode';
 import { getAuth } from 'firebase/auth';
